refactor(project): Rename source files to be valid Python modules
